third_party/WebKit/LayoutTests/editing/spelling/focusing-other-frame.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
</head>
<body>
<pre id="console"></pre>
<iframe id="frame1"></iframe>
<iframe id="frame2"></iframe>

<script>
var frame1 = document.getElementById('frame1');
var testEditable1 = frame1.contentWindow.document.createElement('div');
testEditable1.setAttribute('contentEditable', 'true');
frame1.contentWindow.document.body.appendChild(testEditable1);
var frame2 = document.getElementById('frame2');
var testEditable2 = frame2.contentWindow.document.createElement('div');
testEditable2.setAttribute('contentEditable', 'true');
frame2.contentWindow.document.body.appendChild(testEditable2);

var steps = [
    function() {
        testEditable1.focus();
        frame1.contentWindow.document.execCommand('InsertText', false, 'zz');
    },
    function() {
        testEditable2.focus();
    }
];

var assertions = [
    () => assert_equals(internals.markerCountForNode(testEditable1.childNodes[0], 'spelling'), 0),
    () => assert_equals(internals.markerCountForNode(testEditable1.childNodes[0], 'spelling'), 1)
];

runSpellingTest(steps, assertions,'Text written in an editable in one frame should be spellchecked when focusing other frame')
</script>
</body>
</html>
